pyproject.toml
[project]
name = "boilerplate"
version = "0.1.0"
description = ""
authors = [{ name = "author", email = "author@author.com" }]
dependencies = [
    "fastapi>=0.115.12",
    "motor>=3.3.2",
    "gunicorn>=22.0.0",
    "PyYAML>=6.0.1",
    "uvicorn>=0.25.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.6"
]
requires-python = ">=3.12.2"

[tool.poetry]
package-mode=false

[tool.poetry.group.dev.dependencies]
pytest-asyncio = ">=0.22.0"
mypy = ">=1.8.0"
httpx = ">=0.28.1"

[build-system]
requires = ["poetry-core>=2.0.2"]
build-backend = "poetry.core.masonry.api"
